+++ src/Resource/EnvelopeTemplate.php
@@ -72,4 +72,7 @@ class EnvelopeTemplate extends BaseResource
 
     /** @var Collection<Label> */
     public $labels;
+
+    /** @var Branding|null */
+    public $branding;
 }